Completing RM Restart. Completed Phase 3 of making sure events are logged and restored (mahadev)
git-svn-id: https://svn.apache.org/repos/asf/hadoop/mapreduce/branches/MR-279@1098735 13f79535-47bb-0310-9956-ffa450edef68

+++ yarn/yarn-server/yarn-server-resourcemanager/src/main/java/org/apache/hadoop/yarn/server/resourcemanager/recovery/ZKStore.java
@@ -1,20 +1,20 @@
 /**
-* Licensed to the Apache Software Foundation (ASF) under one
-* or more contributor license agreements.  See the NOTICE file
-* distributed with this work for additional information
-* regarding copyright ownership.  The ASF licenses this file
-* to you under the Apache License, Version 2.0 (the
-* "License"); you may not use this file except in compliance
-* with the License.  You may obtain a copy of the License at
-*
-*     http://www.apache.org/licenses/LICENSE-2.0
-*
-* Unless required by applicable law or agreed to in writing, software
-* distributed under the License is distributed on an "AS IS" BASIS,
-* WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-* See the License for the specific language governing permissions and
-* limitations under the License.
-*/
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
 package org.apache.hadoop.yarn.server.resourcemanager.recovery;
 
 import java.io.IOException;
@@ -68,12 +68,14 @@ public class ZKStore implements Store {
   private static final String ZK_PATH_SEPARATOR = "/";
   private static final String NODE_ID = "nodeid";
   private static final String APP_MASTER = "master";
+  private static final String APP_MASTER_CONTAINER = "mastercontainer";
   private final String ZK_ADDRESS;
   private final int ZK_TIMEOUT;
+  private boolean doneWithRecovery = false;
   
   /** TODO make this generic **/
   private NodeIdPBImpl nodeId = new NodeIdPBImpl();
-  
+
   /**
    * TODO fix this for later to handle all kinds of events 
    * of connection and session events.
@@ -96,25 +98,26 @@ public class ZKStore implements Store {
     );
     this.nodeId.setId(0);
   }
-  
+
   protected Watcher createZKWatcher() {
     return new ZKWatcher();   
   }
-  
+
   private NodeManagerInfoPBImpl createNodeManagerInfo(NodeManager nodeInfo) {
     NodeManagerInfo node = 
       recordFactory.newRecordInstance(NodeManagerInfo.class);
-      node.setNodeAddress(nodeInfo.getNodeAddress());
-      node.setRackName(nodeInfo.getRackName());
-      node.setCapability(nodeInfo.getTotalCapability());
-      node.setUsed(nodeInfo.getUsedResource());
-      node.setNumContainers(nodeInfo.getNumContainers());
-      return (NodeManagerInfoPBImpl)node;
+    node.setNodeAddress(nodeInfo.getNodeAddress());
+    node.setRackName(nodeInfo.getRackName());
+    node.setCapability(nodeInfo.getTotalCapability());
+    node.setUsed(nodeInfo.getUsedResource());
+    node.setNumContainers(nodeInfo.getNumContainers());
+    return (NodeManagerInfoPBImpl)node;
   }
-  
+
   @Override
   public synchronized void storeNode(NodeManager node) throws IOException {
     /** create a storage node and store it in zk **/
+    if (!doneWithRecovery) return;
     NodeManagerInfoPBImpl nodeManagerInfo = createNodeManagerInfo(node);
     byte[] bytes = nodeManagerInfo.getProto().toByteArray();
     try {
@@ -131,6 +134,8 @@ public class ZKStore implements Store {
 
   @Override
   public synchronized void removeNode(NodeManager node) throws IOException {
+    if (!doneWithRecovery) return;
+    
     /** remove a storage node **/
     try {
       zkClient.delete(NODES + Integer.toString(node.getNodeID().getId()), -1);
@@ -141,9 +146,9 @@ public class ZKStore implements Store {
       LOG.info("Keeper exception", ke);
       throw convertToIOException(ke);
     }
-    
+
   }
-  
+
   private static IOException convertToIOException(KeeperException ke) {
     IOException io = new IOException();
     io.setStackTrace(ke.getStackTrace());
@@ -170,48 +175,105 @@ public class ZKStore implements Store {
     String appString = ConverterUtils.toString(containerId.getAppId());
     return appString + "/" + containerId.getId();
   }
-  
-  @Override
-  public synchronized void storeContainer(Container container) throws IOException {
-    ContainerPBImpl containerPBImpl = (ContainerPBImpl) container;
-    try {
-      zkClient.create(APPS + containerPathFromContainerId(container.getId())
-          , containerPBImpl.getProto().toByteArray(), null, CreateMode.PERSISTENT);
-    } catch(InterruptedException ie) {
-      LOG.info("Interrupted", ie);
-      throw new InterruptedIOException(ie.getMessage());
-    } catch(KeeperException ke) {
-      LOG.info("Keeper exception", ke);
-      throw convertToIOException(ke);
+
+  private class ZKApplicationStore implements ApplicationStore {
+    private final ApplicationId applicationId;
+
+    public ZKApplicationStore(ApplicationId applicationId) {
+      this.applicationId = applicationId;
+    }
+
+    @Override
+    public void storeMasterContainer(Container container) throws IOException {
+      if (!doneWithRecovery) return;
+      
+      ContainerPBImpl containerPBImpl = (ContainerPBImpl) container;
+      try {
+        zkClient.setData(APPS + ConverterUtils.toString(container.getId().getAppId()) +
+            ZK_PATH_SEPARATOR + APP_MASTER_CONTAINER
+            , containerPBImpl.getProto().toByteArray(), -1);
+      } catch(InterruptedException ie) {
+        LOG.info("Interrupted", ie);
+        throw new InterruptedIOException(ie.getMessage());
+      } catch(KeeperException ke) {
+        LOG.info("Keeper exception", ke);
+        throw convertToIOException(ke);
+      }
+    }
+    @Override
+    public synchronized void storeContainer(Container container) throws IOException {
+      if (!doneWithRecovery) return;
+      
+      ContainerPBImpl containerPBImpl = (ContainerPBImpl) container;
+      try {
+        zkClient.create(APPS + containerPathFromContainerId(container.getId())
+            , containerPBImpl.getProto().toByteArray(), null, CreateMode.PERSISTENT);
+      } catch(InterruptedException ie) {
+        LOG.info("Interrupted", ie);
+        throw new InterruptedIOException(ie.getMessage());
+      } catch(KeeperException ke) {
+        LOG.info("Keeper exception", ke);
+        throw convertToIOException(ke);
+      }
+    }
+
+    @Override
+    public synchronized void removeContainer(Container container) throws IOException {
+      if (!doneWithRecovery) return;
+      
+      ContainerPBImpl containerPBImpl = (ContainerPBImpl) container;
+      try { 
+        zkClient.delete(APPS + containerPathFromContainerId(container.getId()),
+            -1);
+      } catch(InterruptedException ie) {
+        throw new InterruptedIOException(ie.getMessage());
+      } catch(KeeperException ke) {
+        LOG.info("Keeper exception", ke);
+        throw convertToIOException(ke);
+      }
+    }
+
+    @Override
+    public void updateApplicationState(
+        ApplicationMaster master) throws IOException {
+      if (!doneWithRecovery) return;
+      
+      String appString = APPS + ConverterUtils.toString(applicationId);
+      ApplicationMasterPBImpl masterPBImpl = (ApplicationMasterPBImpl) master;
+      try {
+        zkClient.setData(appString, masterPBImpl.getProto().toByteArray(), -1);
+      } catch(InterruptedException ie) {
+        LOG.info("Interrupted", ie);
+        throw new InterruptedIOException(ie.getMessage());
+      } catch(KeeperException ke) {
+        LOG.info("Keeper exception", ke);
+        throw convertToIOException(ke);
+      }
+    }
+
+    @Override
+    public boolean isLoggable() {
+      return doneWithRecovery;
     }
   }
 
   @Override
-  public synchronized void removeContainer(Container container) throws IOException {
-    ContainerPBImpl containerPBImpl = (ContainerPBImpl) container;
-    try { 
-      zkClient.delete(APPS + containerPathFromContainerId(container.getId()),
-          -1);
-    } catch(InterruptedException ie) {
-      throw new InterruptedIOException(ie.getMessage());
-    } catch(KeeperException ke) {
-      LOG.info("Keeper exception", ke);
-      throw convertToIOException(ke);
-    }
-  }
-
-  @Override
-  public synchronized void storeApplication(ApplicationId application, ApplicationSubmissionContext 
-      context, ApplicationMaster master) throws IOException {
+  public synchronized ApplicationStore createApplicationStore(ApplicationId application, 
+      ApplicationSubmissionContext context) throws IOException {
+    if (!doneWithRecovery) return new ZKApplicationStore(application);
+    
     ApplicationSubmissionContextPBImpl contextPBImpl = (ApplicationSubmissionContextPBImpl) context;
     String appString = APPS + ConverterUtils.toString(application);
-    ApplicationMasterPBImpl masterPBImpl = (ApplicationMasterPBImpl) master;
-    
+   
+    ApplicationMasterPBImpl masterPBImpl = new ApplicationMasterPBImpl();
+    ContainerPBImpl container = new ContainerPBImpl();
     try {
       zkClient.create(appString, contextPBImpl.getProto()
           .toByteArray(), null, CreateMode.PERSISTENT);
       zkClient.create(appString + ZK_PATH_SEPARATOR + APP_MASTER, 
           masterPBImpl.getProto().toByteArray(), null, CreateMode.PERSISTENT);
+      zkClient.create(appString + ZK_PATH_SEPARATOR + APP_MASTER_CONTAINER, 
+          container.getProto().toByteArray(), null, CreateMode.PERSISTENT);
     } catch(InterruptedException ie) {
       LOG.info("Interrupted", ie);
       throw new InterruptedIOException(ie.getMessage());
@@ -219,27 +281,13 @@ public class ZKStore implements Store {
       LOG.info("Keeper exception", ke);
       throw convertToIOException(ke);
     }
+    return new ZKApplicationStore(application);
   }
-  
-  @Override
-  public void updateApplicationState(ApplicationId applicationId,
-      ApplicationMaster master) throws IOException {
-    String appString = APPS + ConverterUtils.toString(applicationId);
-    ApplicationMasterPBImpl masterPBImpl = (ApplicationMasterPBImpl) master;
-    try {
-      zkClient.setData(appString, masterPBImpl.getProto().toByteArray(), -1);
-    } catch(InterruptedException ie) {
-      LOG.info("Interrupted", ie);
-      throw new InterruptedIOException(ie.getMessage());
-    } catch(KeeperException ke) {
-      LOG.info("Keeper exception", ke);
-      throw convertToIOException(ke);
-    }
-  }
-
 
   @Override
   public synchronized void removeApplication(ApplicationId application) throws IOException {
+    if (!doneWithRecovery) return;
+    
     try {
       zkClient.delete(APPS + ConverterUtils.toString(application), -1);
     } catch(InterruptedException ie) {
@@ -250,6 +298,17 @@ public class ZKStore implements Store {
       throw convertToIOException(ke);
     }
   }
+
+  @Override
+  public boolean isLoggable() {
+    return doneWithRecovery;
+  }
+
+  @Override
+  public void doneWithRecovery() {
+    this.doneWithRecovery = true;
+  }
+
   
   @Override
   public synchronized RMState restore() throws IOException {
@@ -257,20 +316,26 @@ public class ZKStore implements Store {
     rmState.load();
     return rmState;
   }  
-  
+
   private class ApplicationInfoImpl implements ApplicationInfo {
     private ApplicationMaster master;
+    private Container masterContainer;
+
     private final ApplicationSubmissionContext context;
     private final List<Container> containers = new ArrayList<Container>();
-    
+
     public ApplicationInfoImpl(ApplicationSubmissionContext context) {
       this.context = context;
     }
-    
+
     public void setApplicationMaster(ApplicationMaster master) {
       this.master = master;
     }
-    
+
+    public void setMasterContainer(Container container) {
+      this.masterContainer = container;
+    }
+
     @Override
     public ApplicationMaster getApplicationMaster() {
       return this.master;
@@ -282,24 +347,29 @@ public class ZKStore implements Store {
     }
 
     @Override
+    public Container getMasterContainer() {
+      return this.masterContainer;
+    }
+
+    @Override
     public List<Container> getContainers() {
       return this.containers;
     }
-    
+
     public void addContainer(Container container) {
       containers.add(container);
     }
   }
-  
+
   private class ZKRMState implements RMState {
     private List<NodeManager> nodeManagers = new ArrayList<NodeManager>();
     private Map<ApplicationId, ApplicationInfo> applications = new 
-      HashMap<ApplicationId, ApplicationInfo>();
-    
+    HashMap<ApplicationId, ApplicationInfo>();
+
     public ZKRMState() {
       LOG.info("Restoring RM state from ZK");
     }
-    
+
     private synchronized List<NodeManagerInfo> listStoredNodes() throws IOException {
       /** get the list of nodes stored in zk **/
       //TODO PB
@@ -322,7 +392,7 @@ public class ZKStore implements Store {
       }
       return nodes;
     }
-    
+
     @Override
     public List<NodeManager> getStoredNodeManagers()  {
       return nodeManagers;
@@ -332,7 +402,7 @@ public class ZKStore implements Store {
     public NodeId getLastLoggedNodeId() {
       return nodeId;
     }
-    
+
     private void readLastNodeId() throws IOException {
       Stat stat = new Stat();
       try {
@@ -346,7 +416,7 @@ public class ZKStore implements Store {
         throw convertToIOException(ke);
       }
     }
-    
+
     private ApplicationInfo getAppInfo(String app) throws IOException {
       ApplicationInfoImpl info = null;
       Stat stat = new Stat();
@@ -363,6 +433,9 @@ public class ZKStore implements Store {
           if (APP_MASTER.equals(child)) {
             master = new ApplicationMasterPBImpl(ApplicationMasterProto.parseFrom(childdata));
             info.setApplicationMaster(master);
+          } else if (APP_MASTER_CONTAINER.equals(child)) {
+            Container masterContainer = new ContainerPBImpl(ContainerProto.parseFrom(data));
+            info.setMasterContainer(masterContainer);
           } else {
             Container container = new ContainerPBImpl(ContainerProto.parseFrom(data));
             info.addContainer(container);
@@ -376,7 +449,7 @@ public class ZKStore implements Store {
       }
       return info;
     }
-    
+
     private void load() throws IOException {
       List<NodeManagerInfo> nodeInfos = listStoredNodes();
       for (NodeManagerInfo node: nodeInfos) {
